apply ts on TokenGroupDTO
